% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binGC.R
\name{binGC}
\alias{binGC}
\title{The genomic sequence was split into multiple sliding windows of the same length and the GC content of each sliding window was calculated.}
\usage{
binGC(filepath, bin = 10000)
}
\arguments{
\item{filepath}{A character vector (of arbitrary length when reading, of length 1 when writing) containing the path(s) to the file(s) to read. Reading files in gzip format (which usually have the '.gz' extension) is supported.

Note that special values like "" or "|cmd" (typically supported by other I/O functions in R) are not supported here.

Note the format of file must be "fasta" or fasta file in gzip.}
\item{bin}{Length of sliding window. Default is 10000.}
}
\value{
Eventually, return a dataframe including the GC content of each sliding window.
}
\description{
The genomic sequence was split into multiple sliding windows of the same length and the GC content of each sliding window was calculated.
}
\references{
\enumerate{
\item Biostrings
\url{https://doi.org/doi:10.18129/B9.bioc.Biostrings}
}

Please, when using quickGenome, cite us using the reference:
\url{https://github.com/nongxinshengxin/quickGenome}
}
\author{
Cy Gao
}
