% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_geneLength.R
\name{calculate_geneLength}
\alias{calculate_geneLength}
\title{The distribution of gene lengths was calculated from gff/gtf.}
\usage{
calculate_geneLength(
  filepath,
  format = "gff",
  type = "genes",
  bin = 500,
  threshold = 4000
)
}
\arguments{
\item{filepath}{Input GFF3 or GTF file. Can be a path to a file, or an URL, or a connection object, or a GFF3File or GTFFile object.}

\item{format}{Format of the input file. Accepted values are: "gff" (the default), or "gtf".}

\item{type}{Counting "genes" (the default) length or "transcripts" length.}

\item{bin}{Length of sliding window. Default is 500}

\item{threshold}{Genes exceeding the threshold length were collapsed together and counted. Default is 4000}
}
\value{
Eventually, return a data frame describing the distribution of gene lengths under a sliding window.
}
\description{
The distribution of gene lengths was calculated from gff/gtf.
}
\references{
\enumerate{
\item GenomicFeatures
\url{https://doi.org/doi:10.18129/B9.bioc.GenomicFeatures}
}

Please, when using quickGenome, cite us using the reference:
\url{https://github.com/nongxinshengxin/quickGenome}
}
\author{
Hm Wang
}
