% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_CDS.R
\name{extract_CDS}
\alias{extract_CDS}
\title{Extraction of CDS sequences and protein sequences}
\usage{
extract_CDS(fasta, gff, outputfile, translation = FALSE)
}
\arguments{
\item{fasta}{A character vector (of arbitrary length when reading, of length 1 when writing) containing the path(s) to the fasta file(s) to read. Reading files in gzip format (which usually have the '.gz' extension) is supported.

Note that special values like "" or "|cmd" (typically supported by other I/O functions in R) are not supported here.

Note the format of file must be "fasta" or fasta file in gzip.}

\item{gff}{A character vector (of arbitrary length when reading, of length 1 when writing) containing the path(s) to the gff file(s) to read.

Note the format of file must be "gff".}

\item{outputfile}{A character vector (of arbitrary length when reading, of length 1 when writing) containing the path(s) to the fasta file(s) to write.}
\item{translation}{TRUE or FALSE (the default). If TRUE, CDS sequence will be translated into a protein sequence.}
}
\description{
Extraction of CDS sequences and protein sequences
}
\references{
\enumerate{
\item Biostrings
\url{https://doi.org/doi:10.18129/B9.bioc.Biostrings}
}

Please, when using quickGenome, cite us using the reference:
\url{https://github.com/nongxinshengxin/quickGenome}
}
\author{
Cy Gao
}
