% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fasta.R
\name{extract_fasta}
\alias{extract_fasta}
\title{Title}
\usage{
extract_fasta(fasta, genelist, outputfile, Negate = FALSE)
}
\arguments{
\item{fasta}{A character vector (of arbitrary length when reading, of length 1 when writing) containing the path(s) to the fasta file(s) to read. Reading files in gzip format (which usually have the '.gz' extension) is supported.

Note that special values like "" or "|cmd" (typically supported by other I/O functions in R) are not supported here.

Note the format of file must be "fasta" or fasta file in gzip.}

\item{genelist}{}

\item{outputfile}{A character vector (of arbitrary length when reading, of length 1 when writing) containing the path(s) to the fasta file(s) to write.}

\item{Negate}{TRUE or FALSE (the default).}

\item{Str}{TRUE or FALSE (the default).}
}
\description{
Title
}
\references{
\enumerate{
\item Biostrings
\url{https://doi.org/doi:10.18129/B9.bioc.Biostrings}
}

Please, when using quickGenome, cite us using the reference:
\url{https://github.com/nongxinshengxin/quickGenome}
}
\author{
Cy Gao
}
