% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_telo.R
\name{find_telo}
\alias{find_telo}
\title{Identification of the location of telomeric repeat sequences at the ends of chromosomes.}
\usage{
find_telo(filepath, telo = "CCCTAA", threshold = 1000, minRepeatNum = 2)
}
\arguments{
\item{filepath}{A character vector (of arbitrary length when reading, of length 1 when writing) containing the path(s) to the file(s) to read or write. Reading files in gzip format (which usually have the '.gz' extension) is supported.

Note that special values like "" or "|cmd" (typically supported by other I/O functions in R) are not supported here.

Note the format of file must be "fasta" or fasta file in gzip.}
\item{telo}{Telomere repeat sequences.}
\item{threshold}{The distance from the end of the chromosome where the repeat sequence occurs. Default is 1000.}
\item{minRepeatNum}{Minimum number of repeats of telomeric repeat sequences. Default is 2.}
}
\value{
Eventually, return a dataframe describing the position of the telomeres at each end of each chromosome.
}
\description{
Identification of the location of telomeric repeat sequences at the ends of chromosomes.
}
\references{
\enumerate{
\item Biostrings
\url{https://doi.org/doi:10.18129/B9.bioc.Biostrings}
}

Please, when using quickGenome, cite us using the reference:
\url{https://github.com/nongxinshengxin/quickGenome}
}
\author{
Cy Gao
}
