% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_basic_Info.R
\name{genome_basic_Info}
\alias{genome_basic_Info}
\title{Calculating basic genomic information.}
\usage{
genome_basic_Info(filepath)
}
\arguments{
\item{filepath}{A character vector (of arbitrary length when reading, of length 1 when writing) containing the path(s) to the file(s) to read. Reading files in gzip format (which usually have the '.gz' extension) is supported.

Note that special values like "" or "|cmd" (typically supported by other I/O functions in R) are not supported here.

Note the format of file must be "fasta" or fasta file in gzip.}
}
\value{
Eventually, the genomic information will be printed on the window.
}
\description{
Calculating basic genomic information, including size, GC content, N50 and so on.
}
\references{
\enumerate{
\item Biostrings
\url{https://doi.org/doi:10.18129/B9.bioc.Biostrings}
}

Please, when using quickGenome, cite us using the reference:
\url{https://github.com/nongxinshengxin/quickGenome}
}
\author{
Cy Gao
}
