% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fasta.R
\name{extract_fasta}
\alias{extract_fasta}
\title{Extracting sequences from fasta files based on sequence names.}
\usage{
extract_fasta(fasta, genelist, outputfile, Negate = FALSE, Str = FALSE)
}
\arguments{
\item{fasta}{A character vector (of arbitrary length when reading, of length 1 when writing) containing the path(s) to the fasta file(s) to read. Reading files in gzip format (which usually have the '.gz' extension) is supported.

Note that special values like "" or "|cmd" (typically supported by other I/O functions in R) are not supported here.

Note the format of file must be "fasta" or fasta file in gzip.}

\item{genelist}{A string containing the path to a file with a list of gene names (or other sequence names).}

\item{outputfile}{A character vector (of arbitrary length when reading, of length 1 when writing) containing the path(s) to the fasta file(s) to write.}

\item{Negate}{TRUE or FALSE (the default). If TRUE, all sequences will be extracted except for those contained in the genelist file.}

\item{Str}{TRUE or FALSE (the default). If TRUE, the genelist parameter requires a string containing the name of a sequence, or a vector containing multiple strings of sequence names.}
}
\description{
Extracting sequences from fasta files based on sequence names.
}
\references{
\enumerate{
\item Biostrings
\url{https://doi.org/doi:10.18129/B9.bioc.Biostrings}
}

Please, when using quickGenome, cite us using the reference:
\url{https://github.com/nongxinshengxin/quickGenome}
}
\author{
Cy Gao
}
